refactor: hooks to remove liquidities
